requirements.txt
asgiref==3.11.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==12.1.0
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
